[Refactor] Refactor single partition desc
Add a new base class SinglePartitionDesc for common variables management,
such as partition name, properties.
